mm,lmn : prev_threshold set to -1 when threshold is added or removed
prev_threshold is a field with the previous current_threshold value.
However, the current_threshold field may have a garbage value at
some point(low_mem_register_event()).

If prev_threshold is set to the value of current_threshold before
current_threshold is explicitly set to a specific value, the
prev_threshold will also have a garbage value.

This happens when low_mem_register_event() is called for the first time.

Anyway, It should be cleared because the threshold corresponding
to the index pointed to by the prev_threshold may not
be valid at this time.